django_api_advance/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
boto3==1.34.29
botocore==1.34.29
captcha==0.5.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
cryptography==42.0.2
dj-database-url==2.1.0
Django==5.0.1
django-crispy-forms==2.1
django-environ==0.11.2
django-otp==1.3.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
mailchimp-marketing==3.0.80
multidict==6.0.4
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.20.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
stripe==8.2.0
twilio==8.12.0
typing_extensions==4.9.0
urllib3==2.0.7
utils==1.0.2
whitenoise==6.6.0
yarl==1.9.4
